requirements.txt
faiss-cpu==1.7.4
ftfy==6.1.1
regex==2023.6.3
tqdm==4.65.0
torch==2.0.1
torchaudio==2.0.2
torchvision==0.15.2
translate==3.6.1
googletrans
wandb==0.14.0
numba==0.57.1
flask==2.3.2
gradio==3.39.0
qdrant-client==1.4.0
httpx
fastdup==1.39
vietocr==0.3.11
fastapi==0.101.1
uvicorn==0.23.2
rapidfuzz==3.3.0
open-clip-torch==2.20.0
flask_cors
colorthief==0.2.1